REGRESSION (Safari 26): Crash in WebProcess due to checkTopOrigin called from WebSWServerConnection::registerServiceWorkerClient with blob urls rdar://160421211 https://bugs.webkit.org/show_bug.cgi?id=298697
Reviewed by Brady Eidson.

The UIProcess is responsible to notify the network process whether a given web process can access cookies for a given top origin.
This logic is used when doing navigation, for instance in WebProcessPool::processForNavigation.
The principle is to get a Site from a request URL and allow the origin.

But the computation of RegistrableDomain for a blob URL was leading to nullOrigin.
This would lead WebProcessPool::processForNavigation to register nullOrigin instead of the actual origin of the blob URL.
This woudl then trigger termination of the web process loading the blob page.

We fix the blob URL logic and add an API test to verify that computation is correct.

* Source/WebCore/fileapi/BlobURL.h:
* Source/WebCore/platform/RegistrableDomain.h:
(WebCore::RegistrableDomain::RegistrableDomain):
(WebCore::RegistrableDomain::registrableDomainFromHost):
* Tools/TestWebKitAPI/Tests/WebCore/RegistrableDomain.cpp:
(TestWebKitAPI::TEST(RegistrableDomain, StringVsURL)):

Canonical link: https://commits.webkit.org/300071@main