[threaded-animation-resolution] accelerated animations do not run https://bugs.webkit.org/show_bug.cgi?id=299016
Reviewed by Anne van Kesteren.

During the bringup of Scroll-driven Animations we refactored the timing code to
require a call to `AnimationEffectTiming::updateComputedProperties()` to resolve
essential timing properties, such as the effect's active duration.

A call to this method is required as an `AcceleratedEffect` is created in the
remote side, so we add one in the remote constructor.

* Source/WebCore/platform/animation/AcceleratedEffect.cpp:
(WebCore::AcceleratedEffect::AcceleratedEffect):

Canonical link: https://commits.webkit.org/300083@main