ReferenceError doesn't include variable name in error message https://bugs.webkit.org/show_bug.cgi?id=275143 rdar://129659300
Reviewed by Keith Miller.

This improves the JSC `ReferenceError` caused by a tdz error by including the variable name in the message. Both Chrome and Firefox do this.
This is achieved by emitting `ExpressionInfo` in the bytecode that corresponds to the identifier. The `ExpressionInfo` is
then used when there is a tdz error. A majority of the changes are fixing tests to follow this format, and the most important changes are
in `NodesCodegen.cpp`

Canonical link: https://commits.webkit.org/300119@main