Array allocation sinking should split allocations into two, an Array allocation and a Butterfly allocation rdar://159207754 https://bugs.webkit.org/show_bug.cgi?id=298606
Reviewed by Yusuke Suzuki.

This patch minorly rearchitects how we do Array allocation sinking in DFG. Previously we tried to model Arrays
in ObjectAllocationSinking as two allocations one where the actual Array was allocated and a "Butterfly"
at each `GetButterfly`. This meant that there was now a reverse data dependency between the GetButterfly and
the Array allocation Nodes. This was a little unintuitive but also meant that any control flow that would
merge two `GetButterfly`s would escape the Array.

This PR simplifies things by more directly representing the heap in ObjectAllocationSinking. There are now
two nodes that get sunk when sinking an Array: NewButterflyWithSize and NewArrayWithButterfly. All the
indexed properties and the Butterfly's location are stored on the NewArrayWithButterfly's Allocation.
The NewButterflyWithSize only contains the Array's length. If we ever wanted to extend Array allocation sinking
to support out of bounds stores we'd have to teach `handleNode` how to deal with writes to multiple allocations.

Originally I had all the indexed properties and the length are stored on the LocalHeap of NewButterflyWithSize and
NewArrayWithButterfly's LocalHeap only contains the butterfly. But I realized that we crash while trying to materialize
a phantom array that contains itself or another newly materialized object that's in a cycle. This is because we have to
materialize the pointer as a PutByVal. We can't emit this PutByVal without the Array since we need to write barrier after
the store.

Overall, this change is perf neutral or maybe a slight progression on JetStream 3.

Canonical link: https://commits.webkit.org/300129@main