[Smart Lists] Availability of Smart Lists need not depend on `_editable` being true https://bugs.webkit.org/show_bug.cgi?id=299056 rdar://160811291
Reviewed by Ryosuke Niwa.

Checking both the web preference and the value of `_editable` is redundant, since the web preference value is sufficient by itself.

Also rename the preference to "availability" instead of "enable" to disambiguate it from when Smart Lists are actually enabled
(which is controlled by the user).

Canonical link: https://commits.webkit.org/300134@main