[GFC] Stop using CanMakeCheckedPtr with UnplacedGridItem https://bugs.webkit.org/show_bug.cgi?id=299069 rdar://160832323
Reviewed by Alan Baradlay.

UnplacedGridItem currently inherits from CanMakeCheckedPtr but seems a
bit unnecessary given how we are currently using it alongside the fact
that it is not designed to be a long-lived object. Instead, we can
achieve the same functionality with two main changes:

1.) Use std::optional<UnplacedGridItem> inside the GridMatrix to
indicate whether an item is within a particular cell or if it is empty.
2.) Add some HashTraits to UnplacedGridItem to allow it to be used
within a HashSet that we use to keep track of which UnplacedGridItems
have been processed when computing ProcessedGridItems.

Canonical link: https://commits.webkit.org/300160@main