Event Timing: assign interactionID to keypress https://bugs.webkit.org/show_bug.cgi?id=298904 rdar://160647204
Reviewed by Ryosuke Niwa.

Assigns an interactionID to keypress events when they are
part of a keydown-keypress sequence. This is not in
accordance to the event timing spec, but is required by
one WPT test (event-timing/interactionid-keypress.html).

Implementing this required deferring the dispatch of keypress
events until the keyup event happens, in a similar manner to
what was already happening for keydown (see the change on
m_pendingKeyDowns).

Matching a keypress to a keydown required an extra function
(KeyboardEvent::keyCodeForKeyDown()) and some change in logic
because the keyCode of keypresses is normally overridden. We now
preserve the original value and do the overriding in
windowsVirtualKeyCode().

Now passing:
* LayoutTests/imported/w3c/web-platform-tests/event-timing/interactionid-keypress.html

Canonical link: https://commits.webkit.org/300198@main