Address Safer CPP warning related to IOServiceMatching() https://bugs.webkit.org/show_bug.cgi?id=299140
Reviewed by Ryosuke Niwa.

The static analysis is not wrong here but there is also no leak because of how the
API is working:
```
Return Value
The matching dictionary created, is returned on success, or zero on failure.
The dictionary is commonly passed to IOServiceGetMatchingServices or
IOServiceAddNotification which will consume a reference, otherwise it should be
released with CFRelease by the caller.
```

I rewrote in a safer way, which should make the static analyser happy.

* Source/WTF/wtf/Compiler.h:
* Source/WebCore/SaferCPPExpectations/RetainPtrCtorAdoptCheckerExpectations:
* Source/WebCore/platform/graphics/mac/GraphicsChecksMac.cpp:
(WebCore::attachToAppleGraphicsControl):

Canonical link: https://commits.webkit.org/300217@main