REGRESSION(300214@main): [macOS Debug] compositing/visible-rect/flipped-preserve-3d.html is a flaky crash https://bugs.webkit.org/show_bug.cgi?id=299195 rdar://160953686
Reviewed by Geoffrey Garen.

When RemoteImageBufferSetProxy gets destroyed, it destroys the RemoteGraphicsContextProxy,
which destroys the DisplayList::Recorder and ends up destroying DrawGlyphsRecorder
which holds a SingleThreadWeakPtr to a Font.

However, we currently let the RemoteImageBufferSetProxy get destroyed off the main thread,
which triggers an assertion in the SingleThreadWeakPtr. We could make the Font support
thread-safe WeakPtr instead but I suspect using SingleThreadWeakPtr was done for performance
reasons. Therefore, I am opting to have the RemoteImageBufferSetProxy get destroyed on the
main thread instead.

* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferSetProxy.h:

Canonical link: https://commits.webkit.org/300289@main