[JSC] Self recursion call in OMG will not need to store callee before calling https://bugs.webkit.org/show_bug.cgi?id=299232 rdar://problem/160990232
Reviewed by Justin Michaud.

When generating a call to itself in OMG, it is guaranteed that the
callee is also OMG since it is calling itself. Since function prologue
stores callee to the CallFrame, the caller side does not need to store
it. This patch detects it and skip storing.

* Source/JavaScriptCore/wasm/WasmOMGIRGenerator.cpp:
(JSC::Wasm::OMGIRGenerator::emitDirectCall):

Canonical link: https://commits.webkit.org/300299@main