Remove ipint argumINT pointer instructions, and extract a constant for ipint alignment. https://bugs.webkit.org/show_bug.cgi?id=298625
Reviewed by Yusuke Suzuki.

This helpful for local builds of the armv7 port, and in general for debug
patches that instrument these opcodes. This way, if the size is ever
too large after instrumentation, it is easy to adjust the alignment globally.

We also add offlineASM support for constexpr align values, to let the
align directives use these constants too.

Canonical link: https://commits.webkit.org/300304@main