[GFC] Inserting items within the grid which fit should not resize the grid. https://bugs.webkit.org/show_bug.cgi?id=299245 rdar://161011615
Reviewed by Alan Baradlay.

Currently, if we try to insert an item which already fits within the
explicit grid, we end up accidentally adding new columns in the grid when
there is no reason to. This is because using Vector::insert will add the
item at the desired index and shuffle the rest of the content over,
increasing the size. The desired outcome is instead to replace the
contents of that cell with the unplaced item we are positioning.

Canonical link: https://commits.webkit.org/300322@main