[GTK][WPE] font-size-adjust:0 and SVG Font failures https://bugs.webkit.org/show_bug.cgi?id=290449
Reviewed by Carlos Garcia Campos.

If the given font lacks the required glyphs, a system fallback font is created.
Previously, the fallback font was initialized with the computed size, which caused
failures in tests using font fallback with font-size-adjust. To fix this, we now
apply font-size-adjust when setting the size of the fallback font.

Test: imported/w3c/web-platform-tests/svg/fonts/zero-font-size-adjust-no-glyphs-font.html
      imported/w3c/web-platform-tests/svg/fonts/zero-font-size-adjust-no-glyphs-zero-height-font.html
      imported/w3c/web-platform-tests/svg/fonts/zero-font-size-adjust.html

* LayoutTests/platform/gtk/TestExpectations:
* LayoutTests/platform/wpe/TestExpectations:
* Source/WebCore/platform/graphics/skia/FontCacheSkia.cpp:
(WebCore::FontCache::systemFallbackForCharacterCluster):

Canonical link: https://commits.webkit.org/300335@main