PLT first meaningful paint metric is sometimes artificially low https://bugs.webkit.org/show_bug.cgi?id=299376 rdar://160884749
Reviewed by Alexey Proskuryakov and Ryosuke Niwa.

The PLT FMP metric is sometimes artifically low. This is because this metric is comes from
committing the CATransaction (in RemoteLayerTreeDrawingAreaProxy::commitLayerTreeTransaction). If
the layer tree commit occurs around the time that we navigate from one page to another, then we
might record a FMP from previous page as part of the FMP of the current page.

To work around this, disallow FMP and other PLT metric updates between the time that we start a
provisional load and when we commit a new load. This seems to work well enough for my testing. If
this continues to be a problem, we might have to look at more invasive workaround like embedding a
document sequence number in to the remote layer tree commit and only emitting FMP if the expected
document sequence numbers match.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::didEndNetworkRequestsForPageLoadTimingTimerFired):
(WebKit::WebPageProxy::didStartProvisionalLoadForFrameShared):
(WebKit::WebPageProxy::didCommitLoadForFrame):
* Source/WebKit/UIProcess/WebPageProxy.h:

Canonical link: https://commits.webkit.org/300408@main