Improve security key PIN re-entry flow https://bugs.webkit.org/show_bug.cgi?id=299119 rdar://160881460
Reviewed by Pascoe.

We need to add a new status update to send to AuthenticationServices to signal
that a PIN was successful. This allows AuthenticationServices to correctly determine
the sequence and timing to present UI in both the success and failure cases.

* Source/WebKit/UIProcess/API/Cocoa/_WKWebAuthenticationPanel.h:
* Source/WebKit/UIProcess/WebAuthentication/AuthenticatorManager.cpp:
(WebKit::AuthenticatorManager::authenticatorStatusUpdated):
* Source/WebKit/UIProcess/WebAuthentication/Cocoa/WebAuthenticationPanelClient.mm:
(WebKit::wkWebAuthenticationPanelUpdate):
* Source/WebKit/UIProcess/WebAuthentication/WebAuthenticationFlags.h:
* Source/WebKit/UIProcess/WebAuthentication/fido/CtapAuthenticator.cpp:
(WebKit::CtapAuthenticator::continueRequestAfterGetPinToken):

Canonical link: https://commits.webkit.org/300411@main