[GFC] Start computing track sizing functions after item placement. https://bugs.webkit.org/show_bug.cgi?id=299320 rdar://problem/161121985
Reviewed by Alan Baradlay.

In preparation for the initial implementation of track sizing, we will
need to compute the track sizing functions after item placement has
occurred. This is because the track sizing functions, along with the
placed items, are needed to perform track sizing. To accomplish this, this
patch does two main things:

1.) Return the number of columns and rows in the implicit grid from
placeGridItems(). placeGridItems will resize the ImplicitGrid as item
placement occurs, so this feels like a natural place to return the
dimensions of the grid as well rather than returning the ImplicitGrid
itself.

2.) Provide the initial code that maps each track to its corresponding
track sizing functions as described in:
https://drafts.csswg.org/css-grid-1/#algo-terms

This is a mostly straightforward mapping from spec text to code, but I
decided not to implement the fit-content case for the max track sizing
function just for simplicity.

Canonical link: https://commits.webkit.org/300417@main