Begin reducing _loadAlternateHTMLString SPI surface area https://bugs.webkit.org/show_bug.cgi?id=299332
Reviewed by Brady Eidson.

When WKWebView._loadAlternateHTMLString is called with a non-empty unreachable URL,
there is code in PolicyChecker::checkNavigationPolicy that bypasses decidePolicyForNavigationAction,
which gives the API no way to pass in a WKWebpagePreferences object.  This PR removes the ability
for the API to pass in arbitrary objects and replaces it with a less powerful WKWebpagePreferences.
This will assist in the transition begun in rdar://160418688, which is already a large change
reducing SPI use.

Test: Tools/TestWebKitAPI/Tests/WebKitCocoa/WebsitePolicies.mm
* Source/WebKit/UIProcess/API/C/WKPage.cpp:
(WKPageLoadAlternateHTMLStringWithUserData):
* Source/WebKit/UIProcess/API/C/WKPage.h:
* Source/WebKit/UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _loadAlternateHTMLString:baseURL:forUnreachableURL:]):
(-[WKWebView _loadAlternateHTMLString:baseURL:forUnreachableURL:withWebpagePreferences:]):
* Source/WebKit/UIProcess/API/Cocoa/WKWebViewPrivate.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::loadAlternateHTML):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WebsitePolicies.mm:

Canonical link: https://commits.webkit.org/300419@main