SVGAElement's rel attribute was never utilised https://bugs.webkit.org/show_bug.cgi?id=298620
Reviewed by Anne van Kesteren.

SVGAElement had rel and relList added in 227520@main.
However, it was never actually used inside of defaultEventHandler()
to change the behaviour of the navigation.

This is now done, alongside the addition of the "opener" value.

SVGAElement now follows the same behaviour as HTMLAnchorElement including the
implicit noopener when target="_blank".

Canonical link: https://commits.webkit.org/300462@main