[css-anchor-position-1] imported/w3c/web-platform-tests/css/css-anchor-position/position-try-cascade-layer-reorder.html is failing https://bugs.webkit.org/show_bug.cgi?id=297159 rdar://157895122
Reviewed by Alan Baradlay.

This test passes in browser but fails in test runner because it is affected by non-overlay scrollbars.
The layer cascading behavior being tested already works as expected.

* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-cascade-layer-reorder-expected.txt:
* Source/WebCore/style/StyleScope.cpp:
(WebCore::Style::Scope::invalidateForLayoutDependencies):

Also take care to call all of the invalidation functions in all cases.

* Source/WebCore/style/StyleTreeResolver.cpp:
(WebCore::Style::scrollContainerSizeForPositionOptions):
(WebCore::Style::TreeResolver::generatePositionOptionsIfNeeded):

When generating the position options save the container scroller (if it is one) size that was used.
Insets with anchor() functions used in the options are resolved against container sides so may
be affected if interleaved layout makes scrollbars appear or disappear.

(WebCore::Style::TreeResolver::tryChoosePositionOption):

If the scroller size has changed when it is time to choose an option then re-generate the options.
This is guaranteed to make progress because we save the last tried option and continue from that.

* Source/WebCore/style/StyleTreeResolver.h:

Canonical link: https://commits.webkit.org/300465@main